refactor: extract CommandDeleteDialog
